-update script proper name -added Ronin market qoute endpoint
